update keys for consistency.
In the docs, we are walking the user through and have suggested names for the
internal, private, protected and public buckets.  This just reminds them to use
the same ones during the walkthrough.

stackname and stacknamenodash are also updated to be more inline with examples
for iam and deployer CF stack configuration.