Add zip archive validation method to zipLambda
We need to be sure there isn't an invalid package file left over
from a previous aborted kes run - as the hash is generated and
written as the filestream is opened in kes.utils.zip, and
kes is capable of async calling process.exit, it's possible
to have truncated/0 byte zip files.  This validator ensures
that the file is a legitimate zip archive in addition to
a matching hash filename.