Change eventSourceMapping's StartingPosition LATEST => AT_TIMESTAMP
When the StartingPosition of a rule was set at LATEST to read the most recent
messages, adding a message to teh stream did not result in a triggered
stack-kinesisConsumer lambda event.

This change now will have the eventsourcemapping read all of the events as of
the mapping creation time.  (I think this may cause problems with long lasting
streams, if a consumer is added to an event source mapping it may read all of
the events on the stream, but I'm not 100% sure if this is actually a problem)