Wait for test SF execution out of possibly many. [it]
When we destroy/recreate the Kinesis rule in PopulateProvidersCollections.js,
We basically re-initialize the kinesis shard starting position to the
TRIM_HORIZON. This means that when your test is run, there may be any number of
previous events in the stream.

This means that when a new event triggers the kinesisConsumer as a result of
your test code, your particular execution may not be the latest execution.

This will grab the last 20 executions and starting with the most recent, check
the payloads of the SfSnsReport steps for your tests recordIdentifier. and
return that execution to the caller.