Use eslint-airbnb-base instead of eslint-airbnb
From the eslint-airbnb readme: "If you don't need React, see
eslint-config-airbnb-base."

This removes the dependencies on eslint-plugin-react and
eslint-plugin-jsx-a11y.

This also resulted in a number of new types of warnings being generated,
almost all of which were stylistic in nature.  Those that are in the api
package have been addressed here.