Update kes override to only build/push lambdas once
Update the kes override to only build/push the lambdas on the first
iteration of the nested stack set that needs it.   For the parent
we're going to make the unhappy compromise of recalculating the
bucket/hash/etc for the lambdas that are presumed to have been built
in the child stack.