Update templates/build structure to explicitly build all lambdas
Non-workflow lambdas were being grouped with a lambda in the
lambdas.yml file (InRegionS3Policy) which was resulting in the
internal API lambdas being grouped and avoiding any build/upload
issues if they weren't evaluated for build in the processing step.

Should someone attempt to deploy without InRegionS3Policy, or add a
lambda to the non-workflow api lambdas, the current PR will fail
to build/upload the lambda code.    This PR remedies that.