Clarifies Kinesis test failure cases.
Adds a new expectation to 'outputs the record' to catch the case where the
failure is not the type we expect when we put a bad record on the kinesis
stream.  This will fail while a bug occurs like in CUMULUS-948.

Checks the return results from getRecords and fails more clearly if the results
are empty.