Inject CMR metadata into api collection list() results.
Right now, list is hitting elasticsearch and we haven't stored the concept ids
when possible, so we have to search each record that comes out and update it
with the concept-id if it exists.

concept-id injection tests.

remove logs add docs

removes useless test.

remove requires and eslint fixes.

renamed injectCMRmetadata=>injectConceptId.

That's all it's doing right now, so if we expand it we can rename it later.

simplify error handling.

Smaller fixtures

hack at stubbing getConceptId;

refactor injectConceptId allowing conceptIdFunction to be passed in

This just lets us replace the getConceptId with a different function for
testing.

jsdocs update

remove unused imports