Remove table creation/deletion from test.
In conversation with the team, we're going with an approach that runs
the migration to create the tables in CI, and relys on local devs to
setup their database manually in development.    Individual records
should be generated such that they're assured to be unique to avoid test collisions in dev.