Switch back to just bucket.name
Before in move-granules, the file was updated with the full bucket object
file.bucket = "public"
became
file.bucket = {type: 'public', name: 'public-bucket-name'}

This goes back and uses the stack's bucket config to get the type by name in a
bunch of places and it needs to be cleaned up.

But I wanted this in.