Test for metadata that is currently being chopped off
When we move granules, the code is currently replacing all URLS with only the
updated files.  This will delete any metadata in OnlineAccessUrls with the list
of new files' urls.

The correct thing will be to merge the new files URL objects into the original
urls the metadata initially had discarding updated ones.