Change eventual consistency strategy.
Before, I was waiting in the test and assuming the api would also be
consistent, that isn't the case, so this commit removes all of the waiting for
statistics in the test, and just calls the API over and over until we get the
result we expect.

This is after a lot of hand testing and loops to see that even if the metric is
found from the test, the api may still be behind that node.