Update kes handler to ifEqualsString.
Something in the merging of yaml files makes it difficult to guess whether the
value in the interpolated config will be a boolean or string.  This just casts
both sides to string and makes the comparison explicit.  The ifEquals command
is still available from the parent kes deployment (and the one I changed was
never being used).