Cumulus 800 add api gateway logging (#1055)
* Force ApiGateway logging to cloudwatch

* Add logging to apiConfig in config.yml

* Adds CloudwatchLogsToKinesisRole to iam template

This role will allow putRecord on to the cross account "Destination" kinesis
stream.

* Creates Log Groups and enables Subscription filters.

* removes cloudwatchLogsToKinesisRole for shared log destination.

Before we had cross account destinations, we had  a kinesis stream of our own,
and we needed a role to write to it.  now that is handled by the shared log destination.

* Adds lambda option logToSharedDestination

This adds a AWS::Logs::SubscriptionFilter and logGroup for the lambda that will
submit logs to a configured AWS::Logs::Destination.

* Add CF/kes config for lambda's to log on shared Destination

This allows any lambda function to use the logToSharedDestination key and have
the cloudwatch logs shipped to the configured shared Logs::Destination.

* Have to delete the dev stage by hand.

But you also have to update the deployments to new values so they don't collide
with the earlier values.

* eslint :bell:

* More lint :bell:

* Instructions on api-gateway-logging.

* adds policy document to readme.

* And new environment variable to bamboo

* removes and ignores i18n/en.json

This file is generated on each build and shouldn't be committed.

* Info for enabling CloudWatch logs from API Gateway

* Move delete-stage from example/scripts to deployment package.

A user who has installed from the template doesn't have access to the example
directory, so include the delete-stage script where they can access it.

* Enable AccessLogSetting

* rewords changelog.

* Update config to allow over-riding default deployment

Chose to set logging to a shared destination as OFF because it's only for NASA
deployments.

* Update kes handler to ifEqualsString.

Something in the merging of yaml files makes it difficult to guess whether the
value in the interpolated config will be a boolean or string.  This just casts
both sides to string and makes the comparison explicit.  The ifEquals command
is still available from the parent kes deployment (and the one I changed was
never being used).

* Doc Additions

* merge keys sharedLogDestinationArn and logToSharedDestination

Rather than two keys, now we use logToSharedDestination and set the value to a
writable ARN rather than true, and getting the value from from sharedLogDestinationArn.

* Further docs updates. to move info out of CHANGELOG.

* Docs updates

* Docs Changes.

* get rid of now unused ifEqualsString

* PR Feedback.