Adds ByLogStream to subscriptionFilters
I think terraform has a bug where if you don't specify a `distribution`
attribute, it defaults to "ByLogStream", and then forgets so that a second
terraform apply shows you a change from "ByLogStream" => null,
but that apply doesn't do it. so you always have a change with terraform apply
and it never stablizes even with no changes.