Adds default level sit_cma_layer to config.yml
We can't share lambda layers yet across accounts so need to have separate
versions for the cma_layer in each environment.

We add a sandbox_cma_layer and a sit_cma_layer

Then we override the default cma_layer with sandbox_cma_layer and override the
sit deployment specific cma_layer with the sit_cma_layer.