Change PUT of a collection to replace collection
HTTP PUT of a collection now replaces the collection rather than
updating it, making it conform to the standard. PATCH should be used for
updates rather than replacements.

Also, PUT now responds with 400 rather than 404 when URI path contains
a name and/or version that does not match the name and/or version of the
collection payload, and added unit tests for both 400 and 404 responses.

CUMULUS-1308