Change PUT of a provider to replace provider
HTTP PUT of a provider now replaces the provider rather than updating
it, making it conform to the standard. PATCH should be used for updates
rather than replacements.

Also, PUT now responds with 400 when URI path contains an ID that does
not match the ID of the provider payload, and added unit tests for both
400 and 404 responses.

CUMULUS-1308