Change PUT of a rule to replace rule
HTTP PUT of a rule now replaces the rule rather than updating
it, making it conform to the standard. PATCH should be used for updates
rather than replacements.

Also, PUT now responds with 400 when URI path contains a name that does
not match the name of the rule payload, and added unit tests for both
400 and 404 responses.

CUMULUS-1308