Include dependency library template in api lambda.
saml2 library has a problem:
https://github.com/Clever/saml2/issues/150

But larger than that problem is the way a sublibrary is used
xml-encryption uses templates that are read from disk.
Here's a PR that would fix that and solve this problem:
https://github.com/auth0/node-xml-encryption/pull/40/files

Long story short, when we build a webpacked app/index.js, the xml-encryption
templates must be located next to the webpacked data.