Remove unused packages
While taking a look at some of the package deprecation warnings being reported by npm, I noticed that there were 1) a lot of packages being installed that were not being used and 2) a number of packages being installed as dependencies that should have been dev dependencies.

This commit removes the unused deps, and moves things to devDependencies when appropriate.