CUMULUS-1308 Handle deletion of fields via the API (#1198)
* Test for replace, not modify, on collection update

The HTTP PUT operation for a collection should perform an update/replace
rather than an update/modify (merge) of the existing value, which is how
HTTP PUT is intended to behave (update/modify should use POST).  This
test now checks for full replacement of a collection upon PUT.

CUMULUS-1308

* Change PUT of a collection to replace collection

HTTP PUT of a collection now replaces the collection rather than
updating it, making it conform to the standard. PATCH should be used for
updates rather than replacements.

Also, PUT now responds with 400 rather than 404 when URI path contains
a name and/or version that does not match the name and/or version of the
collection payload, and added unit tests for both 400 and 404 responses.

CUMULUS-1308

* Change PUT of a provider to replace provider

HTTP PUT of a provider now replaces the provider rather than updating
it, making it conform to the standard. PATCH should be used for updates
rather than replacements.

Also, PUT now responds with 400 when URI path contains an ID that does
not match the ID of the provider payload, and added unit tests for both
400 and 404 responses.

CUMULUS-1308

* Change PUT of a rule to replace rule

HTTP PUT of a rule now replaces the rule rather than updating
it, making it conform to the standard. PATCH should be used for updates
rather than replacements.

Also, PUT now responds with 400 when URI path contains a name that does
not match the name of the rule payload, and added unit tests for both
400 and 404 responses.

CUMULUS-1308

* Add CHANGELOG entry

CUMULUS-1308

* Add Chuck's CI stack

CUMULUS-1308

* Change API Rule tests to pass complete object

Rule tests now pass a complete rule object on update via the API since
the API now expects a full object, not a partial object.

CUMULUS-1308

* Use update, not create, for Rule model update

CUMULUS-1308

* fix update params for enabling rule test in snsRuleSpec

* add specific error when passing rule.arn to enable a disabled rule

* add unit test for enabling a disabled rule and passing rule.arn

* Fix test to expect correct error message

CUMULUS-1308

* Update test to pass full rule object when rerunning rule

CUMULUS-1308

* Fix eslint errors

CUMULUS-1308

* WIP

* Add/delete collection during rule tests

CUMULUS-1308

* Whitelist https-proxy-agent module

* Pass complete collection object to API on update in tests

CUMULUS-1308