CUMULUS-1657: Fix docs (#1364)
* fix path to interfaces diagram

* minor fixes to deployment README

* another docs fix

* fix typo

* Update docs/deployment/README.md

Co-Authored-By: Jonathan Kovarik <kovarik@nsidc.org>

* CUMULUS-1572: Fix granule discovery for files without file type config (#1279)

* Fix granule discovery for files without file type config

A granule is now included in discovery results even when none of its
files has a matching file type in the associated collection
configuration.  Previously, if all files for a granule were unmatched by
a file type configuration, the granule was excluded from the discovery
results.  Now, instead, such a granule is included in the discovery
results, but with an empty file list.

Added corresponding unit and integration tests for this fix.

Also addressed 3 ESLint warnings for JSDoc violations on the
`addUrlPathToGranuleFiles` and `addUniqueGranuleFilePathToGranuleFiles`
functions in `example/spec/helpers/granuleUtils.js`, along with some
refactoring of those functions to eliminate unnecessary (excess) object
cloning.

CUMULUS-1572

* Define property to ignore collection files config

Added support for the `boolean` property `ignoreFilesConfig` at both the
task config level for the `DiscoverGranules` task, as well as at the
collection configuration level for overriding the task-level value.
By default, the property is `false`, meaning that during granule
discovery, a file for a granule is ingested only when the name of the
file matches one of the regexes in the collection's `files` config,
which is the current behavior.

However, setting this property to `true` on the `DiscoverGranules` task
will cause all `files` configs for all collections to be ignored, such
that all files for a granule will be ingested (because there will be no
filename checks against the regexes in the `files` config).

However, it is possible to override the property value set on the task
by setting the property on individual collections.

CUMULUS-1572

* Rename property and change precedence.

Renamed property `ignoreFilesConfig` to `ignoreFilesConfigForDiscovery`
to better reflect its narrow usage. Also, changed precedence such that
setting this property on a task overrides the property value on all
collections.

CUMULUS-1572

* Document Collection attribute `ignoreFilesConfigForDiscovery`

CUMULUS-1572

* Make test names unique

CUMULUS-1572

* Add missing asyncs/awaits

Added missing asyncs/awaits that were erroneoulsy removed because of
misleading VSCode hints that were caused by incorrect JSDoc comments.
Also corrected the JSDoc comments to prevent the misleading hints.

CUMULUS=1572

* Add missing awaits

CUMULUS-1572

* Allow deploying ES to multiple subnets (#1359)

* Allow deploying ES to multiple subnets

* Apply suggestions from code review

Co-Authored-By: Mark Boyd <markdavidboyd@gmail.com>

* Fix bad merge

* fix bad links in docs README

* update docs

* minor update