Add ability to retain database when running localApi
Adds new options to the cli to allow the dashboard to start up the api without
wiping the dynamoDB and Elasticsearch tables.  This allows the user to seed
with whatever data they want before running the api.  The defaults are the same
as they were before and nothing else should be affected.