Cumulus 1102 run cumulus api locally in dashboard (#1415)
* Fix path for workflow

This works when running from different directories. Like the
`cumulus-api reset-tables` call.

* Update testAuth to use JWT

Use a JWT in test authorization.

* Nuke randomstring in favor of useful strings.

* Adds eraseDatastack function to serve.js and exports is.

This is probably a step on a refactor because I don't think this should live
here.  But I'm trying to make sure this spike will work with the dashboard when
it uses the api as a package.

* refactor createDBRecords to use serveUtils

Just moved some helpers into new file so they can be used by dashboard.

* Expert helpers through serve for now.

* Add ability to retain database when running localApi

Adds new options to the cli to allow the dashboard to start up the api without
wiping the dynamoDB and Elasticsearch tables.  This allows the user to seed
with whatever data they want before running the api.  The defaults are the same
as they were before and nothing else should be affected.

* Relocates workflows, adds reconcile-reports

Updates localrun, but not 100% sure this won't break cumulus integration tests.

* Fix path for reconcilation-reports.

* Add reconciliaion reports fixtures to localAPI

* remove overkill logging.

* shame :bell:

* is it a local only problem

* is it a local only problem x2

* :facepalm:

* Not sure this is allowed

* This seems correct.

* use it all

* goodnight

* Don't try to use stepfunction API in testing.

* update CHANGELOG

* can't use that like I thought.

* update CHANGELOG

* Enable stepfunction api on localstack.

* regression in localstack bucket names.

Looks like there was a regression in valid localstack bucketnames.

* Can't even have a capitalized bucket letter.

* Remove incorrect comments.

* Adds logs back for debugging.

* reorder CHANGELOG and unCamelCase bucketNames

* Remove exports of helper functions.

* Update testAuth to handle delete and refresh.

* Implements PR feedback suggestions.

Refactors single letter variable names.
Separates concerns with eraseLocalElasticsearch.
fixes bug in promise array.

* PR Feedback continued

Remove unused function fetchElasticDefaults.
Removes unnecessary default params.

* Fix lint

* Apply suggestions from code review

Better wording

Co-Authored-By: Mark Boyd <markdavidboyd@gmail.com>

* Refactors server tests defaults into separate package.

* Use correct values from getESClientAndIndex

* wait for ES

* adds NO_DEPRECATION_WARNINGS variable

For testing the deprecation warnings get in the way, if you set this variable
to anything it will prevent those warnings from showing up.

* Pull errors from @cumulus/errors instead of @cumulus/common/errors.

* Use alpha version of cumulus/api

Co-authored-by: Mark Boyd <markdavidboyd@gmail.com>