Adding "errors" to the FilesInCumulus report.
I'm not sure if this is the right way to handle this, but we need to do
something.  The integration tests are creating valid Granule.File objects that
aren't parsed correctly (they have no 'key' property).  I can parse source, but
I think that just leads us down a weird path.