cumulus-1961/replace dynamodb with ES in parts of reconciliation report creation (#1737)
* Adds start[end]Timestamp parameters to reconcilation report

Also renaming of endpoint functions for clarity.

* Don't await for simple array.

* small cleanup refactor.

* Adds ability to configure server to use ES tunneled to AWS.

Need to set LOCAL_ES_HOST to localhost and LOCAL_ES_HOST_PROTOCOL to https and
you are able to run the api locally but hit AWS for ES and lambdas.

* Return list of files.

I don't think this is what we actually want.
We need to return something like what dynamo does.

* Use ES instead of dynamo for internal file check vs s3.

* replace dynamo with ES for createReconciliationReportForBucket

uses elasticsearch files queue instead of dynamo db to pull data to compare to S3.

* fixup! Merge remote-tracking branch 'origin/master' into mhs/CUMULUS-1961/add-start-end-params-to-reconciliation-report-creation

* Fix first test.

* Pass start/end Timestamps to lambda.

* Add logging on ES errors.

Just log the error and toss.

* Ignore unmapped as type 'keyword'

Update tests for randomID

* Fix Second Test.

* Pass all reconciliation tests.

* Adds tests for esFileSearchQueue

* Lint fix.

* spy on elasticsearch.Client.scroll

* add more spy tests.

* Update changelog

* missed lint

* Update payload in tests

* Fix es/search tests.

LOCAL_ES_HOST became something that determines that can override the aws
production hosts.  This prevents that in tests.

* Fix spec

Guess at the problem.

* Create elasticOnly records for tests.

* Cleanup after tests.

* Tweaks to names and comments.

* local elasticsearch client accesses tunnelled Elasticsearc cluster.

* Allow full sort params to be passed in via queryStringParams

This just alllows a user to provied
sortParams: {sort: <anything> }}
and the query will provide
"sort: <anything>" in the response.

The default values for sort_by and order still work unless sortParams is
provided.

* Adds ScrollSearch class for reconciliation reports

* Update ESFileSearchQueue tests.

* Pass params to ESScrollSearch correctly.

* Replace DynamoDb for Elasticsearch for collections differences.

* Use ES for collection reconciliation

* Deletes test data from DyamoDB

Only for reconciliation reports that are migrated.

* Adds ESCollectionGranuleAQueue

This is for iterating over granules by collectionId in
reconciliationReportForGranules

* Lint :bell:

* Test Cleanup

* rename esFileSearchQueue => esFileQueue

* remove todo date for lint?

* Pin aws-elasticsearch-connector to 8.2.0

* Try to debug bamboo deployment.

* Adding "errors" to the FilesInCumulus report.

I'm not sure if this is the right way to handle this, but we need to do
something.  The integration tests are creating valid Granule.File objects that
aren't parsed correctly (they have no 'key' property).  I can parse source, but
I think that just leads us down a weird path.

* Lint :bell:

* REVERT reconcilationReportForBuckets to dynamoDB Granules.Cache

This reverts reconcilationReportForBuckets back to use the Granules file
cache.

This just until we can reindex and upgrade Elasticsearch to use a new query.

* Deleting the granule too many times?

* Update Changelog removing ref to createReconciliationReportForBucket

* removes S3Errors and ESErrors.

Should have been removed before.

* force ES refresh before reconcilation reports.

Not 100% sure this is a good idea, but also not sure this fixes the problem.

* Address feedback: only support granule for this class.

* PR typo.

* Catch errors in beforeAll to make debugging easier.

This lets you not worry about errors when the beforeAll is what fails.

* Removes params from _buildSearch()

That shouldn't be there.

* Spy on client's scroll function.

* Refactors tests to use common helpers.

* remove refreshing of ES indices.

This wasn't the problem.

* Wait for granule to be removed from ES.

* Lint :bell:

* Fix tests:  why?

* Ignore helper files during testing

* Tests esSearchQueue

* Removes sort order from tests for esSearchQueue.

* Adds tests for esCollectionGranulesQueue.

Also renames grans => granules

* Pulling in Marc's change for #1774

* renaming variables.

* Separate out the failed test start from the first test.

* DEBUGGING INFORMATION.  REVERT THIS. FOR REAL.

This is shotgun logging thrown at the reconcilation report.
It adds information on all ok files and granules to the report.

Additionally, it does not clean up your stack data if a test fails when you run
the integration tests.

* Fix granules API client list querying

* Update Changelog

* Revert "DEBUGGING INFORMATION.  REVERT THIS. FOR REAL."

This reverts commit d8d2e51ab38619c07da7d3f5b46734e66e348f04.

* Test fix

Co-authored-by: Lauren Frederick <lpfreder@mbp-lpfreder.edn.ecs.nasa.gov>