Do packaging in subshell, and only if there exists a package.json
The bootstrap process was hanging when it tried to run the second half of the
bootstrap (npm run package).

The example/package.json had a command that would run into an infinite loop
without printing output to the screen if there was a subdirectory under the
example/lambda without a package.json file.  This is due to running `npm run
package` in a subdir that kicked itself off again.

This change outputs the current directory being packaged, and then only runs
the package script if there is a package.json file.

This should prevent problems with switching branches when working on sample lambdas.