Update reconcilation report to delete collection/deps
Reconciliation report can fail on cleanup/a number of other
instances.    When this occurs, it becomes impossible for the
beforeAll setup to succeed.    Of the two options, one being
pre-delete the collection and all it's dependencies and/or upserting
the collection, I've opted to do both as PRs for now for evaluation.