Ensure the default esIndex is created properly before test.
These lines were added because on the first run of the test against localstack,
the defaultIndexName does not exist but another test later on creates the index
which causes the setup of this to fail on subsequent runs.  it succeeded in CI
because we use a new es container on each run.