Update db package to use executions table for write constraint
This commit updates the granule upsert logic to evaluate
running/queued status writes for existance of
an execution and it's status from the executions table instead of the
linking table.  This is due to the linking record not being written
before the granule causing an out-of-order write edge case when the
running *and* completed granule are being written at the same tiem.