Update data-migration2 to handle empty values for granule.execution (#2997)
* Update data-migration2 to handle empty values for granule.execution

This update modifies the data-migration2 script to allow for dynamo
granule.execution values of empty string/''/etc.   This is in response
to CUMULUS-2961/a user report that they have executions returning from
dynamo with an empty string value, which was resulting in the
execution lookup query to fail.   This is presumptively due to the
inclusion of the granule API endpoints that allow direct user
mainupulation of data given the granules are recent and the
event-driven write code doesn't allow for a situation where the
execution is missing from dynamo.

* Fix logging output

* Update CHANGELOG

* Apply suggestions from code review