Release 11.1.3 backport (#3001)
* CUMULUS-2929 backport

* CUMULUS-2863

* CUMULUS-2959

* update to v11.1.3

* update package api doc

* update changelog

* CUMULUS-2966: Added extractPath operation and support of nested string replacement to url_path (#3007)

* Added extractPath operation and nested string replacement to url_path in the collection configuration

* update changelog

* update unit test descriptions

Co-authored-by: Jennifer Tran <12633533+botanical@users.noreply.github.com>

* Update data-migration2 to handle empty values for granule.execution (#2997)

* Update data-migration2 to handle empty values for granule.execution

This update modifies the data-migration2 script to allow for dynamo
granule.execution values of empty string/''/etc.   This is in response
to CUMULUS-2961/a user report that they have executions returning from
dynamo with an empty string value, which was resulting in the
execution lookup query to fail.   This is presumptively due to the
inclusion of the granule API endpoints that allow direct user
mainupulation of data given the granules are recent and the
event-driven write code doesn't allow for a situation where the
execution is missing from dynamo.

* Fix logging output

* Update CHANGELOG

* Apply suggestions from code review

* update changelog

Co-authored-by: Jennifer Tran <12633533+botanical@users.noreply.github.com>
Co-authored-by: Jonathan Kovarik <kovarik@nsidc.org>