Encryption & Decryption:  decrypt is imported and used, but you have not provided the implementation for this function. Ensure that it handles errors correctly. In the getPrivateKey() method, KMS.decryptBase64String(fetchedKey) is used, but it appears the result isn't awaited. If decryptBase64String is asynchronous, you need to use await.
Encryption & Decryption:

decrypt is imported and used, but you have not provided the implementation for this function. Ensure that it handles errors correctly.
In the getPrivateKey() method, KMS.decryptBase64String(fetchedKey) is used, but it appears the result isn't awaited. If decryptBase64String is asynchronous, you need to use await.