fix sftp client not connected bug
Running the basic usage code in docs does not work and always throw "Client not connected" error. Upon inspection I found that a `connect` method is defined that connects to an SFTP server. But that method is not called anywhere. This is fixed by calling `connect` in the constructor that then opens an SFTP connection for subsequent use.