CUMULUS-3682 -- Update @cumulus/message.Granule to handle missing values in CMR metadata (#3865)
* Update @cumulus/message.Granule to handle missing values in CMR metadata

This commit adds logic in convertDateToISOStringSettingNull to

- set '' as an expected null if encountered in any of the granule
temporal fields handled by the Granules methods.
- Handle explicit undefined values by returning them rather than
attempting to date convert them (the error case for this ticket)

This update should be compatible with api granule write logic in that
it's preserving the intent of implicit undefined values while allowing
for explicit undefineds that weren't originally part of the expected
typing coming from the CMR metadata spec.

* Update CHANGELOG